Fix type of $timeout parameter of JsonClient::query
Php::Type::Numeric is an integer, not a float.
See:
- https://github.com/yaroslavche/phptdlib/blob/0.0.10/tdlib.cpp#L57
- https://github.com/CopernicaMarketingSoftware/PHP-CPP/blob/v2.2.0/include/type.h#L25

+++ stubs/TDLib/JsonClient.php
@@ -14,11 +14,11 @@ class JsonClient extends BaseJsonClient {
 
     /**
      * @param string $query - json-encoded string request
-     * @param float|null $timeout
+     * @param int|null $timeout
      * @return string - json-encoded library response
      * @throws \Exception
      */
-    public function query(string $query, float $timeout = null):string
+    public function query(string $query, int $timeout = null):string
     {
     }
